Move list initialization out of model.Update
Something is off here which causes a "hang" at the list.New when using
the `--reference` flag. (I think there's a better explanation but I'm
not finding it yet.) For now, move the list initialization to `run` and
then set the actual contents of the models when the messages are
received, which seemingly fix the hang.

Fixes #72.